jumphost-mcp/pyproject.toml
[project]
name = "srefix-jumphost-mcp"
version = "0.1.0"
description = "MCP server for SSH-via-jumphost remote command execution with safety gating"
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.0.0",
    "PyYAML>=6.0",
]

[project.scripts]
srefix-jumphost = "jumphost_mcp.server:run"

[build-system]
requires = ["setuptools>=64"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
